+++ src/core/custom-counter/custom-counter.styled.ts
@@ -1,7 +1,9 @@
+import styled from 'styled-components/native'
+import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons'
 import { COLOR } from '@core/colors/colors.constants'
 import { RESPONSIVE } from '@core/responsive/responsive'
 import { TEXT_SIZE } from '@core/text/text.constants'
-import styled from 'styled-components/native'
+import { getThemeProp } from '@core/styled/theme'
 
 export const CustomCounterContainer = styled.View`
   flex-direction: row;
@@ -25,7 +27,7 @@ export const CustomCounterTextContainer = styled.View`
 
 export const CustomCounterText = styled.Text`
   font-size: ${TEXT_SIZE.M}px;
-  color: ${COLOR.BLACK};
+  color: ${getThemeProp('textPrimary')};
 `
 
 export const CustomCounterButtonContainer = styled.TouchableOpacity`
@@ -36,3 +38,21 @@ export const CustomCounterButtonContainer = styled.TouchableOpacity`
   border: 1px solid ${COLOR.DIM_GREY_LIGHTER};
   border-radius: 10px;
 `
+
+export const CustomCounterLabel = styled(MaterialCommunityIcons).attrs(props => ({
+  name: 'format-letter-spacing',
+  color: getThemeProp('textPrimary')(props),
+  size: TEXT_SIZE.L,
+}))``
+
+export const CustomCounterMinusIcon = styled(MaterialCommunityIcons).attrs(props => ({
+  name: 'minus',
+  color: getThemeProp('textPrimary')(props),
+  size: TEXT_SIZE.XXL,
+}))``
+
+export const CustomCounterPlusIcon = styled(MaterialCommunityIcons).attrs(props => ({
+  name: 'plus',
+  color: getThemeProp('textPrimary')(props),
+  size: TEXT_SIZE.XXL,
+}))``