+++ src/core/letters-slider/components/letter-slider-label/letter-slider-label.styled.ts
@@ -1,11 +1,11 @@
 import styled from 'styled-components/native'
 
 export const LetterSliderLabel = styled.View`
-  width: 40px;
+  width: 50px;
   height: 25px;
   position: absolute;
   bottom: -55px;
-  left: -20px;
+  left: -27.5px;
   border-radius: 10px;
   justify-content: center;
   align-items: center;