+++ src/core/letters-slider/letter-slider.styled.ts
@@ -1,8 +1,8 @@
 import styled from 'styled-components/native'
 import RangeSlider from 'rn-range-slider'
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons'
+import { getThemeProp } from '@core/styled/theme'
 import { TEXT_SIZE } from '../text/text.constants'
-import { COLOR } from '../colors/colors.constants'
 
 export const LetterSliderBottomLabelBar = styled.View`
   flex-direction: row;
@@ -17,7 +17,7 @@ export const LetterSliderTopLabelBar = styled.View`
 
 export const LetterSliderBottomLabel = styled.Text`
   font-size: ${TEXT_SIZE.M}px;
-  color: ${COLOR.BLACK};
+  color: ${getThemeProp('textPrimary')};
   font-weight: bold;
   margin-left: 10px;
   margin-right: 5px;
@@ -27,10 +27,10 @@ export const LetterSlider = styled(RangeSlider)`
   width: 100%;
 `
 
-export const LetterSliderLengthIcon = styled(MaterialCommunityIcons).attrs({
+export const LetterSliderLengthIcon = styled(MaterialCommunityIcons).attrs(props => ({
   name: 'format-letter-spacing',
-  color: COLOR.BLACK,
+  color: getThemeProp('textPrimary')(props),
   size: TEXT_SIZE.L,
-})`
+}))`
   padding-top: 3px;
 `