+++ src/dashboard/components/word-details-modal/word-details-modal.styled.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components/native'
 import { TEXT_SIZE } from '@core/text/text.constants'
-import { COLOR } from '@core/colors/colors.constants'
+import { getThemeProp } from '@core/styled/theme'
 
 export const WordDetailsModalContainer = styled.View`
   width: 100%;
@@ -13,7 +13,7 @@ export const WordDetailsModalContainer = styled.View`
 
 export const WordDetailsHeadlineText = styled.Text`
   font-size: ${TEXT_SIZE.XL}px;
-  color: ${COLOR.BLACK};
+  color: ${getThemeProp('textPrimary')};
   font-weight: bold;
   text-transform: uppercase;
   text-align: center;
@@ -22,7 +22,7 @@ export const WordDetailsHeadlineText = styled.Text`
 
 export const WordDetailsDefinitionText = styled.Text`
   font-size: ${TEXT_SIZE.S}px;
-  color: ${COLOR.BLACK};
+  color: ${getThemeProp('textPrimary')};
   font-weight: bold;
   text-align: center;
   margin-bottom: 10px;